refactor(snippets): introduce SnippetsRenderer
