+++ snippets/src/update.spec.ts
@@ -1,6 +1,7 @@
 import 'jasmine';
 
 import { SnippetRepository } from './SnippetRepository';
+import { SnippetsRenderer } from './SnippetsRenderer';
 import { update } from './update';
 
 const markdown = `
@@ -16,38 +17,18 @@ I had some old content
 <!--END_DOCUSAURUS_CODE_TABS-->
 `.trimStart();
 
-
-const snippetWithoutExample = `
-let Foo = {
-  x: 1
-}
-in Foo
-`.trimStart();
-
-const snippetWithExample = `
-let Foo = {
-  x: 1
-}
-let example = {
-  y: 1
-}
-in Foo
-`.trimStart();
-
 const repo: Partial<SnippetRepository> = {
   findSnippet: async (id) => {
     switch (id) {
       case 'a':
-        return { id: 'a', content: snippetWithoutExample };
+        return { id: 'a', content: 'foo' };
       case 'b':
-        return { id: 'b', content: snippetWithExample };
+        return { id: 'b', content: 'bar' };
       default:
         return null;
     }
   }
 };
-
-const fence = '```';
 
 const expected = `
 # First
@@ -55,49 +36,34 @@ const expected = `
 first text
 
 <!--snippet:a-->
-
-The following configuration options are available at \`a\`:
-<!--DOCUSAURUS_CODE_TABS-->
-<!--Dhall Type-->
-${fence}haskell
-let Foo = {
-  x: 1
-}
-${fence}
+foo
 <!--END_DOCUSAURUS_CODE_TABS-->
 
 <!--snippet:b-->
-
-The following configuration options are available at \`b\`:
-<!--DOCUSAURUS_CODE_TABS-->
-<!--Dhall Type-->
-${fence}haskell
-let Foo = {
-  x: 1
-}
-${fence}
-<!--Example-->
-${fence}haskell
-let example = {
-  y: 1
-}
-${fence}
+bar
 <!--END_DOCUSAURUS_CODE_TABS-->
 `.trimStart();
 
+const renderer: Partial<SnippetsRenderer> = {
+  render: (s) => `
+<!--snippet:${s.id}-->
+${s.content}
+<!--END_DOCUSAURUS_CODE_TABS-->`.trimStart()
+};
 
 describe('update', () => {
 
-  const repoMock = repo as SnippetRepository;
+  const stubRepo = repo as SnippetRepository;
+  const stubRenderer = renderer as SnippetsRenderer;
 
   it('updates multiple snippets in one go', async () => {
-    const result = await update(markdown, repoMock);
+    const result = await update(markdown, stubRepo, stubRenderer);
 
     expect(result).toEqual(expected);
   });
 
   it('updates are idempotent', async () => {
-    const result = await update(expected, repoMock);
+    const result = await update(expected, stubRepo, stubRenderer);
 
     expect(result).toEqual(expected);
   });
@@ -109,6 +75,6 @@ describe('update', () => {
     <!--END_DOCUSAURUS_CODE_TABS-->
     `;
 
-    await expectAsync(update(markdownWithInvalidReference, repoMock)).toBeRejectedWithError();
+    await expectAsync(update(markdownWithInvalidReference, stubRepo, stubRenderer)).toBeRejectedWithError();
   });
 });
